Adjust to tiff-4.4.0 output
tiffinfo of tiff-4.4.0 added a directory index and -M option. That
broke t/92_tiffinfo.t.

This patch adds CurrentDirectory() method to the Perl module and
prints the directory index by tiffinfo.pl. Of that breaks
compatibility with the previous tiff library versions.

This patch adjusts t/92_tiffinfo.t to ignore the new -M option. The
option is used to refuse loading large images.

https://rt.cpan.org/Ticket/Display.html?id=143153