Handle TIFFTAG_XPOSITION and TIFFTAG_YPOSITION
ImageMagick somwehere after 7.1.1.4 and before 7.1.1.9 started to
define a position in the output TIFF images. Those are used in tests and
the unxepected field caused a failure like this:

    $ prove -b t/1.t
    t/1.t .. 1/50
    #   Failed test 'PrintDirectory'
    #   at t/1.t line 122.
    #          got: '466'
    #     expected: '449'

    #   Failed test 'tiffcmp'
    #   at t/1.t line 157.
    #          got: 'XPosition tag appears only in /tmp/MkyBKzzTV7/test.tif
    # '
    #     expected: ''
    # Looks like you failed 2 tests of 50.
    t/1.t .. Dubious, test returned 2 (wstat 512, 0x200)
    Failed 2/50 subtests

This patch adds TIFFTAG_XPOSITION and TIFFTAG_YPOSITION to
Graphics::TIFF close to already existing TIFFTAG_XRESOLUTION and
TIFFTAG_YRESOLUTION. Then the patch adapts the tests.

Signed-off-by: Petr Písař <ppisar@redhat.com>
Signed-off-by: Jeffrey Ratcliffe <jffry@posteo.net>